@@ -20,10 +20,9 @@ package clickhouse
 import (
 	"context"
 	"io"
-	"io/ioutil"
 )
 
-func (h *httpConnect) asyncInsert(ctx context.Context, query string, wait bool) error {
+func (h *httpConnect) asyncInsert(ctx context.Context, query string, wait bool, args ...any) error {
 
 	options := queryOptions(ctx)
 	options.settings["async_insert"] = 1
@@ -31,11 +30,19 @@ func (h *httpConnect) asyncInsert(ctx context.Context, query string, wait bool) 
 	if wait {
 		options.settings["wait_for_async_insert"] = 1
 	}
+	if len(args) > 0 {
+		var err error
+		query, err = bindQueryOrAppendParameters(true, &options, query, h.location, args...)
+		if err != nil {
+			return err
+		}
+	}
+
 	res, err := h.sendQuery(ctx, query, &options, h.headers)
 	if res != nil {
 		defer res.Body.Close()
 		// we don't care about result, so just discard it to reuse connection
-		_, _ = io.Copy(ioutil.Discard, res.Body)
+		_, _ = io.Copy(io.Discard, res.Body)
 	}
 
 	return err
