@@ -19,62 +19,53 @@ package clickhouse
 
 import (
 	"context"
-	"errors"
 	"fmt"
+	"io"
+	"slices"
+
 	"github.com/ClickHouse/clickhouse-go/v2/lib/column"
 	"github.com/ClickHouse/clickhouse-go/v2/lib/driver"
 	"github.com/ClickHouse/clickhouse-go/v2/lib/proto"
-	"io"
-	"io/ioutil"
-	"regexp"
-	"strings"
 )
 
-// \x60 represents a backtick
-var httpInsertRe = regexp.MustCompile(`(?i)^INSERT INTO\s+\x60?([\w.^\(]+)\x60?\s*(\([^\)]*\))?`)
+// release is ignored, because http used by std with empty release function.
+// Also opts ignored because all options unused in http batch.
+func (h *httpConnect) prepareBatch(ctx context.Context, query string, opts driver.PrepareBatchOptions, release func(*connect, error), acquire func(context.Context) (*connect, error)) (driver.Batch, error) {
+	query, tableName, queryColumns, err := extractNormalizedInsertQueryAndColumns(query)
+	if err != nil {
+		return nil, err
+	}
 
-// release is ignored, because http used by std with empty release function
-func (h *httpConnect) prepareBatch(ctx context.Context, query string, release func(*connect, error), acquire func(context.Context) (*connect, error)) (driver.Batch, error) {
-	matches := httpInsertRe.FindStringSubmatch(query)
-	if len(matches) < 3 {
-		return nil, errors.New("cannot get table name from query")
-	}
-	tableName := matches[1]
-	var rColumns []string
-	if matches[2] != "" {
-		colMatch := strings.TrimSuffix(strings.TrimPrefix(matches[2], "("), ")")
-		rColumns = strings.Split(colMatch, ",")
-		for i := range rColumns {
-			rColumns[i] = strings.TrimSpace(rColumns[i])
-		}
-	}
-	query = "INSERT INTO " + tableName + " FORMAT Native"
-	queryTableSchema := "DESCRIBE TABLE " + tableName
-	r, err := h.query(ctx, release, queryTableSchema)
+	describeTableQuery := fmt.Sprintf("DESCRIBE TABLE %s", tableName)
+	r, err := h.query(ctx, release, describeTableQuery)
 	if err != nil {
 		return nil, err
 	}
 
 	block := &proto.Block{}
 
-	// get Table columns and types
 	columns := make(map[string]string)
 	var colNames []string
 	for r.Next() {
 		var (
-			colName string
-			colType string
-			ignore  string
+			colName      string
+			colType      string
+			default_type string
+			ignore       string
 		)
 
-		if err = r.Scan(&colName, &colType, &ignore, &ignore, &ignore, &ignore, &ignore); err != nil {
+		if err = r.Scan(&colName, &colType, &default_type, &ignore, &ignore, &ignore, &ignore); err != nil {
 			return nil, err
+		}
+		// these column types cannot be specified in INSERT queries
+		if default_type == "MATERIALIZED" || default_type == "ALIAS" {
+			continue
 		}
 		colNames = append(colNames, colName)
 		columns[colName] = colType
 	}
 
-	switch len(rColumns) {
+	switch len(queryColumns) {
 	case 0:
 		for _, colName := range colNames {
 			if err = block.AddColumn(colName, column.Type(columns[colName])); err != nil {
@@ -83,7 +74,7 @@ func (h *httpConnect) prepareBatch(ctx context.Context, query string, release fu
 		}
 	default:
 		// user has requested specific columns so only include these
-		for _, colName := range rColumns {
+		for _, colName := range queryColumns {
 			if colType, ok := columns[colName]; ok {
 				if err = block.AddColumn(colName, column.Type(colType)); err != nil {
 					return nil, err
@@ -193,6 +184,7 @@ func (b *httpBatch) Send() (err error) {
 		headers["Content-Encoding"] = b.conn.compression.String()
 	case CompressionZSTD, CompressionLZ4:
 		options.settings["decompress"] = "1"
+		options.settings["compress"] = "1"
 	}
 
 	go func() {
@@ -223,10 +215,18 @@ func (b *httpBatch) Send() (err error) {
 	if res != nil {
 		defer res.Body.Close()
 		// we don't care about result, so just discard it to reuse connection
-		_, _ = io.Copy(ioutil.Discard, res.Body)
+		_, _ = io.Copy(io.Discard, res.Body)
 	}
 
 	return err
+}
+
+func (b *httpBatch) Rows() int {
+	return b.block.Rows()
+}
+
+func (b *httpBatch) Columns() []column.Interface {
+	return slices.Clone(b.block.Columns)
 }
 
 var _ driver.Batch = (*httpBatch)(nil)
