@@ -22,10 +22,10 @@ import (
 	"crypto/x509"
 	"database/sql"
 	"fmt"
-	"github.com/ClickHouse/clickhouse-go/v2"
-	"io/ioutil"
 	"os"
 	"path"
+
+	"github.com/ClickHouse/clickhouse-go/v2"
 )
 
 func ConnectSSL() error {
@@ -38,7 +38,7 @@ func ConnectSSL() error {
 		return err
 	}
 	t := &tls.Config{}
-	caCert, err := ioutil.ReadFile(path.Join(cwd, "../../tests/resources/CAroot.crt"))
+	caCert, err := os.ReadFile(path.Join(cwd, "../../tests/resources/CAroot.crt"))
 	if err != nil {
 		return err
 	}
