@@ -25,6 +25,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/ClickHouse/clickhouse-go/v2/lib/column"
 	"github.com/ClickHouse/clickhouse-go/v2/lib/driver"
 )
 
@@ -58,8 +59,10 @@ func DateNamed(name string, value time.Time, scale TimeUnit) driver.NamedDateVal
 	}
 }
 
-var bindNumericRe = regexp.MustCompile(`\$[0-9]+`)
-var bindPositionalRe = regexp.MustCompile(`[^\\][?]`)
+var (
+	bindNumericRe    = regexp.MustCompile(`\$[0-9]+`)
+	bindPositionalRe = regexp.MustCompile(`[^\\][?]`)
+)
 
 func bind(tz *time.Location, query string, args ...any) (string, error) {
 	if len(args) == 0 {
@@ -109,41 +112,64 @@ func checkAllNamedArguments(args ...any) (bool, error) {
 	return haveNamed, nil
 }
 
-var bindPositionCharRe = regexp.MustCompile(`[?]`)
-
 func bindPositional(tz *time.Location, query string, args ...any) (_ string, err error) {
 	var (
-		unbind = make(map[int]struct{})
-		params = make([]string, len(args))
+		lastMatchIndex = -1 // Position of previous match for copying
+		argIndex       = 0  // Index for the argument at current position
+		buf            = make([]byte, 0, len(query))
+		unbindCount    = 0 // Number of positional arguments that couldn't be matched
 	)
-	for i, v := range args {
-		if fn, ok := v.(std_driver.Valuer); ok {
-			if v, err = fn.Value(); err != nil {
-				return "", nil
+
+	for i := 0; i < len(query); i++ {
+		// It's fine looping through the query string as bytes, because the (fixed) characters we're looking for
+		// are in the ASCII range to won't take up more than one byte.
+		if query[i] == '?' {
+			if i > 0 && query[i-1] == '\\' {
+				// Copy all previous index to here characters
+				buf = append(buf, query[lastMatchIndex+1:i-1]...)
+				buf = append(buf, '?')
+			} else {
+				// Copy all previous index to here characters
+				buf = append(buf, query[lastMatchIndex+1:i]...)
+
+				// Append the argument value
+				if argIndex < len(args) {
+					v := args[argIndex]
+					if fn, ok := v.(std_driver.Valuer); ok {
+						if v, err = fn.Value(); err != nil {
+							return "", nil
+						}
+					}
+
+					value, err := format(tz, Seconds, v)
+					if err != nil {
+						return "", err
+					}
+
+					buf = append(buf, value...)
+					argIndex++
+				} else {
+					unbindCount++
+				}
 			}
-		}
-		params[i], err = format(tz, Seconds, v)
-		if err != nil {
-			return "", err
+
+			lastMatchIndex = i
 		}
 	}
-	i := 0
-	query = bindPositionalRe.ReplaceAllStringFunc(query, func(n string) string {
-		if i >= len(params) {
-			unbind[i] = struct{}{}
-			return ""
-		}
-		val := params[i]
-		i++
-		return bindPositionCharRe.ReplaceAllStringFunc(n, func(m string) string {
-			return val
-		})
-	})
-	for param := range unbind {
-		return "", fmt.Errorf("have no arg for param ? at position %d", param)
+
+	// If there were no replacements, quick return without copying the string
+	if lastMatchIndex < 0 {
+		return query, nil
 	}
-	// replace \? escape sequence
-	return strings.ReplaceAll(query, "\\?", "?"), nil
+
+	// Append the remainder
+	buf = append(buf, query[lastMatchIndex+1:]...)
+
+	if unbindCount > 0 {
+		return "", fmt.Errorf("have no arg for param ? at last %d positions", unbindCount)
+	}
+
+	return string(buf), nil
 }
 
 func bindNumeric(tz *time.Location, query string, args ...any) (_ string, err error) {
@@ -221,6 +247,12 @@ func bindNamed(tz *time.Location, query string, args ...any) (_ string, err erro
 func formatTime(tz *time.Location, scale TimeUnit, value time.Time) (string, error) {
 	switch value.Location().String() {
 	case "Local", "":
+		// It's required to pass timestamp as string due to decimal overflow for higher precision,
+		// but zero-value string "toDateTime('0')" will be not parsed by ClickHouse.
+		if value.Unix() == 0 {
+			return "toDateTime(0)", nil
+		}
+
 		switch scale {
 		case Seconds:
 			return fmt.Sprintf("toDateTime('%d')", value.Unix()), nil
@@ -238,14 +270,16 @@ func formatTime(tz *time.Location, scale TimeUnit, value time.Time) (string, err
 		return fmt.Sprintf("toDateTime64('%s', %d)", value.Format(fmt.Sprintf("2006-01-02 15:04:05.%0*d", int(scale*3), 0)), int(scale*3)), nil
 	}
 	if scale == Seconds {
-		return value.Format(fmt.Sprintf("toDateTime('2006-01-02 15:04:05', '%s')", value.Location().String())), nil
+		return fmt.Sprintf("toDateTime('%s', '%s')", value.Format("2006-01-02 15:04:05"), value.Location().String()), nil
 	}
 	return fmt.Sprintf("toDateTime64('%s', %d, '%s')", value.Format(fmt.Sprintf("2006-01-02 15:04:05.%0*d", int(scale*3), 0)), int(scale*3), value.Location().String()), nil
 }
 
+var stringQuoteReplacer = strings.NewReplacer(`\`, `\\`, `'`, `\'`)
+
 func format(tz *time.Location, scale TimeUnit, v any) (string, error) {
 	quote := func(v string) string {
-		return "'" + strings.NewReplacer(`\`, `\\`, `'`, `\'`).Replace(v) + "'"
+		return "'" + stringQuoteReplacer.Replace(v) + "'"
 	}
 	switch v := v.(type) {
 	case nil:
@@ -278,7 +312,45 @@ func format(tz *time.Location, scale TimeUnit, v any) (string, error) {
 		}
 		return fmt.Sprintf("[%s]", val), nil
 	case fmt.Stringer:
+		if v := reflect.ValueOf(v); v.Kind() == reflect.Pointer &&
+			v.IsNil() &&
+			v.Type().Elem().Implements(reflect.TypeOf((*fmt.Stringer)(nil)).Elem()) {
+			return "NULL", nil
+		}
 		return quote(v.String()), nil
+	case column.OrderedMap:
+		values := make([]string, 0)
+		for key := range v.Keys() {
+			name, err := format(tz, scale, key)
+			if err != nil {
+				return "", err
+			}
+			value, _ := v.Get(key)
+			val, err := format(tz, scale, value)
+			if err != nil {
+				return "", err
+			}
+			values = append(values, fmt.Sprintf("%s, %s", name, val))
+		}
+
+		return "map(" + strings.Join(values, ", ") + ")", nil
+	case column.IterableOrderedMap:
+		values := make([]string, 0)
+		iter := v.Iterator()
+		for iter.Next() {
+			key, value := iter.Key(), iter.Value()
+			name, err := format(tz, scale, key)
+			if err != nil {
+				return "", err
+			}
+			val, err := format(tz, scale, value)
+			if err != nil {
+				return "", err
+			}
+			values = append(values, fmt.Sprintf("%s, %s", name, val))
+		}
+
+		return "map(" + strings.Join(values, ", ") + ")", nil
 	}
 	switch v := reflect.ValueOf(v); v.Kind() {
 	case reflect.String:
