@@ -22,7 +22,7 @@ import (
 	"fmt"
 	"os"
 	"regexp"
-	"strings"
+	"slices"
 	"time"
 
 	"github.com/pkg/errors"
@@ -32,27 +32,15 @@ import (
 	"github.com/ClickHouse/clickhouse-go/v2/lib/proto"
 )
 
-var splitInsertRe = regexp.MustCompile(`(?i)\sVALUES\s*\(`)
-var columnMatch = regexp.MustCompile(`.*\((?P<Columns>.+)\)$`)
+var insertMatch = regexp.MustCompile(`(?i)(INSERT\s+INTO\s+[^( ]+(?:\s*\([^()]*(?:\([^()]*\)[^()]*)*\))?)(?:\s*VALUES)?`)
+var columnMatch = regexp.MustCompile(`INSERT INTO .+\s\((?P<Columns>.+)\)$`)
 
-func (c *connect) prepareBatch(ctx context.Context, query string, release func(*connect, error), acquire func(context.Context) (*connect, error)) (driver.Batch, error) {
-	//defer func() {
-	//	if err := recover(); err != nil {
-	//		fmt.Printf("panic occurred on %d:\n", c.num)
-	//	}
-	//}()
-	query = splitInsertRe.Split(query, -1)[0]
-	colMatch := columnMatch.FindStringSubmatch(query)
-	var columns []string
-	if len(colMatch) == 2 {
-		columns = strings.Split(colMatch[1], ",")
-		for i := range columns {
-			columns[i] = strings.Trim(strings.TrimSpace(columns[i]), "`")
-		}
+func (c *connect) prepareBatch(ctx context.Context, query string, opts driver.PrepareBatchOptions, release func(*connect, error), acquire func(context.Context) (*connect, error)) (driver.Batch, error) {
+	query, _, queryColumns, verr := extractNormalizedInsertQueryAndColumns(query)
+	if verr != nil {
+		return nil, verr
 	}
-	if !strings.HasSuffix(strings.TrimSpace(strings.ToUpper(query)), "VALUES") {
-		query += " VALUES"
-	}
+
 	options := queryOptions(ctx)
 	if deadline, ok := ctx.Deadline(); ok {
 		c.conn.SetDeadline(deadline)
@@ -71,30 +59,41 @@ func (c *connect) prepareBatch(ctx context.Context, query string, release func(*
 		return nil, err
 	}
 	// resort batch to specified columns
-	if err = block.SortColumns(columns); err != nil {
+	if err = block.SortColumns(queryColumns); err != nil {
 		return nil, err
 	}
-	return &batch{
-		ctx:         ctx,
-		conn:        c,
-		block:       block,
-		released:    false,
-		connRelease: release,
-		connAcquire: acquire,
-		onProcess:   onProcess,
-	}, nil
+
+	b := &batch{
+		ctx:          ctx,
+		query:        query,
+		conn:         c,
+		block:        block,
+		released:     false,
+		connRelease:  release,
+		connAcquire:  acquire,
+		onProcess:    onProcess,
+		closeOnFlush: opts.CloseOnFlush,
+	}
+
+	if opts.ReleaseConnection {
+		b.release(b.closeQuery())
+	}
+
+	return b, nil
 }
 
 type batch struct {
-	err         error
-	ctx         context.Context
-	conn        *connect
-	sent        bool
-	released    bool
-	block       *proto.Block
-	connRelease func(*connect, error)
-	connAcquire func(context.Context) (*connect, error)
-	onProcess   *onProcess
+	err          error
+	ctx          context.Context
+	query        string
+	conn         *connect
+	sent         bool // sent signalize that batch is send to ClickHouse.
+	released     bool // released signalize that conn was returned to pool and can't be used.
+	closeOnFlush bool // closeOnFlush signalize that batch should close query and release conn when use Flush
+	block        *proto.Block
+	connRelease  func(*connect, error)
+	connAcquire  func(context.Context) (*connect, error)
+	onProcess    *onProcess
 }
 
 func (b *batch) release(err error) {
@@ -122,11 +121,48 @@ func (b *batch) Append(v ...any) error {
 	if b.err != nil {
 		return b.err
 	}
+
+	if len(v) > 0 {
+		if r, ok := v[0].(*rows); ok {
+			return b.appendRowsBlocks(r)
+		}
+	}
+
 	if err := b.block.Append(v...); err != nil {
 		b.err = errors.Wrap(ErrBatchInvalid, err.Error())
 		b.release(err)
 		return err
 	}
+	return nil
+}
+
+// appendRowsBlocks is an experimental feature that allows rows blocks be appended directly to the batch.
+// This API is not stable and may be changed in the future.
+// See: tests/batch_block_test.go
+func (b *batch) appendRowsBlocks(r *rows) error {
+	var lastReadLock *proto.Block
+	var blockNum int
+
+	for r.Next() {
+		if lastReadLock == nil { // make sure the first block is logged
+			b.conn.debugf("[batch.appendRowsBlocks] blockNum = %d", blockNum)
+		}
+
+		// rows.Next() will read the next block from the server only if the current block is empty
+		// only if new block is available we should flush the current block
+		// the last block will be handled by the batch.Send() method
+		if lastReadLock != nil && lastReadLock != r.block {
+			if err := b.Flush(); err != nil {
+				return err
+			}
+			blockNum++
+			b.conn.debugf("[batch.appendRowsBlocks] blockNum = %d", blockNum)
+		}
+
+		b.block = r.block
+		lastReadLock = r.block
+	}
+
 	return nil
 }
 
@@ -147,12 +183,15 @@ func (b *batch) IsSent() bool {
 
 func (b *batch) Column(idx int) driver.BatchColumn {
 	if len(b.block.Columns) <= idx {
-		b.release(nil)
+		err := &OpError{
+			Op:  "batch.Column",
+			Err: fmt.Errorf("invalid column index %d", idx),
+		}
+
+		b.release(err)
+
 		return &batchColumn{
-			err: &OpError{
-				Op:  "batch.Column",
-				Err: fmt.Errorf("invalid column index %d", idx),
-			},
+			err: err,
 		}
 	}
 	return &batchColumn{
@@ -166,43 +205,71 @@ func (b *batch) Column(idx int) driver.BatchColumn {
 }
 
 func (b *batch) Send() (err error) {
+	stopCW := contextWatchdog(b.ctx, func() {
+		// close TCP connection on context cancel. There is no other way simple way to interrupt underlying operations.
+		// as verified in the test, this is safe to do and cleanups resources later on
+		if b.conn != nil {
+			_ = b.conn.conn.Close()
+		}
+	})
+
 	defer func() {
+		stopCW()
 		b.sent = true
 		b.release(err)
 	}()
-	if b.sent {
-		return b.retry()
-	}
 	if b.err != nil {
 		return b.err
 	}
-	if b.block.Rows() != 0 {
-		if err = b.conn.sendData(b.block, ""); err != nil {
+	if b.sent || b.released {
+		if err = b.resetConnection(); err != nil {
 			return err
 		}
 	}
-	if err = b.conn.sendData(&proto.Block{}, ""); err != nil {
-		return err
+	if b.block.Rows() != 0 {
+		if err = b.conn.sendData(b.block, ""); err != nil {
+			// there might be an error caused by context cancellation
+			// in this case we should return context error instead of net.OpError
+			if ctxErr := b.ctx.Err(); ctxErr != nil {
+				return ctxErr
+			}
+
+			return err
+		}
 	}
-	if err = b.conn.process(b.ctx, b.onProcess); err != nil {
+	if err = b.closeQuery(); err != nil {
 		return err
 	}
 	return nil
 }
 
-func (b *batch) retry() (err error) {
-	// exit early if Send() hasn't been attepted
-	if !b.sent {
-		return ErrBatchNotSent
-	}
-
+func (b *batch) resetConnection() (err error) {
 	// acquire a new conn
 	if b.conn, err = b.connAcquire(b.ctx); err != nil {
 		return err
 	}
-	b.sent = false
-	b.released = false
-	return b.Send()
+
+	defer func() {
+		b.released = false
+	}()
+
+	options := queryOptions(b.ctx)
+	if deadline, ok := b.ctx.Deadline(); ok {
+		b.conn.conn.SetDeadline(deadline)
+		defer b.conn.conn.SetDeadline(time.Time{})
+	}
+
+	if err = b.conn.sendQuery(b.query, &options); err != nil {
+		b.release(err)
+		return err
+	}
+
+	if _, err = b.conn.firstBlock(b.ctx, b.onProcess); err != nil {
+		b.release(err)
+		return err
+	}
+
+	return nil
 }
 
 func (b *batch) Flush() error {
@@ -212,12 +279,40 @@ func (b *batch) Flush() error {
 	if b.err != nil {
 		return b.err
 	}
+	if b.released {
+		if err := b.resetConnection(); err != nil {
+			return err
+		}
+	}
 	if b.block.Rows() != 0 {
 		if err := b.conn.sendData(b.block, ""); err != nil {
 			return err
 		}
+		if b.closeOnFlush {
+			b.release(b.closeQuery())
+		}
 	}
 	b.block.Reset()
+	return nil
+}
+
+func (b *batch) Rows() int {
+	return b.block.Rows()
+}
+
+func (b *batch) Columns() []column.Interface {
+	return slices.Clone(b.block.Columns)
+}
+
+func (b *batch) closeQuery() error {
+	if err := b.conn.sendData(&proto.Block{}, ""); err != nil {
+		return err
+	}
+
+	if err := b.conn.process(b.ctx, b.onProcess); err != nil {
+		return err
+	}
+
 	return nil
 }
 
@@ -229,12 +324,11 @@ type batchColumn struct {
 }
 
 func (b *batchColumn) Append(v any) (err error) {
+	if b.err != nil {
+		return b.err
+	}
 	if b.batch.IsSent() {
 		return ErrBatchAlreadySent
-	}
-	if b.err != nil {
-		b.release(b.err)
-		return b.err
 	}
 	if _, err = b.column.Append(v); err != nil {
 		b.release(err)
@@ -244,12 +338,11 @@ func (b *batchColumn) Append(v any) (err error) {
 }
 
 func (b *batchColumn) AppendRow(v any) (err error) {
+	if b.err != nil {
+		return b.err
+	}
 	if b.batch.IsSent() {
 		return ErrBatchAlreadySent
-	}
-	if b.err != nil {
-		b.release(b.err)
-		return b.err
 	}
 	if err = b.column.AppendRow(v); err != nil {
 		b.release(err)
