@@ -23,40 +23,36 @@ import (
 )
 
 func MultiHostVersion() error {
+	return multiHostVersion(nil)
+}
+
+func MultiHostRoundRobinVersion() error {
+	connOpenStrategy := clickhouse.ConnOpenRoundRobin
+	return multiHostVersion(&connOpenStrategy)
+}
+
+func MultiHostRandomVersion() error {
+	connOpenStrategy := clickhouse.ConnOpenRandom
+	return multiHostVersion(&connOpenStrategy)
+}
+
+func multiHostVersion(connOpenStrategy *clickhouse.ConnOpenStrategy) error {
 	env, err := GetNativeTestEnvironment()
 	if err != nil {
 		return err
 	}
-	conn, err := clickhouse.Open(&clickhouse.Options{
+	options := clickhouse.Options{
 		Addr: []string{"127.0.0.1:9001", "127.0.0.1:9002", fmt.Sprintf("%s:%d", env.Host, env.Port)},
 		Auth: clickhouse.Auth{
 			Database: env.Database,
 			Username: env.Username,
 			Password: env.Password,
 		},
-	})
-	if err != nil {
-		return err
 	}
-	v, err := conn.ServerVersion()
-	if err != nil {
-		return err
+	if connOpenStrategy != nil {
+		options.ConnOpenStrategy = *connOpenStrategy
 	}
-	fmt.Println(v.String())
-	return nil
-}
-
-func MultiHostRoundRobinVersion() error {
-	env, err := GetNativeTestEnvironment()
-	conn, err := clickhouse.Open(&clickhouse.Options{
-		Addr:             []string{"127.0.0.1:9001", "127.0.0.1:9002", fmt.Sprintf("%s:%d", env.Host, env.Port)},
-		ConnOpenStrategy: clickhouse.ConnOpenRoundRobin,
-		Auth: clickhouse.Auth{
-			Database: env.Database,
-			Username: env.Username,
-			Password: env.Password,
-		},
-	})
+	conn, err := clickhouse.Open(&options)
 	if err != nil {
 		return err
 	}
