@@ -20,10 +20,11 @@ package std
 import (
 	"context"
 	"fmt"
+	"time"
+
 	"github.com/ClickHouse/clickhouse-go/v2"
 	clickhouse_tests "github.com/ClickHouse/clickhouse-go/v2/tests/std"
 	"github.com/google/uuid"
-	"time"
 )
 
 func UseContext() error {
@@ -36,17 +37,14 @@ func UseContext() error {
 	}
 	// we can use context to pass settings to a specific API call
 	ctx := clickhouse.Context(context.Background(), clickhouse.WithSettings(clickhouse.Settings{
-		"allow_experimental_object_type": "1",
+		"async_insert": "1",
 	}))
-	conn.ExecContext(ctx, "DROP TABLE IF EXISTS example")
-	// to create a JSON column we need allow_experimental_object_type=1
-	if _, err = conn.ExecContext(ctx, `
-		CREATE TABLE example (
-				Col1 JSON
-			) 
-			Engine Memory
-		`); err != nil {
-		return err
+	var settingValue bool
+	if err := conn.QueryRowContext(ctx, "SELECT getSetting('async_insert')").Scan(&settingValue); err != nil {
+		return fmt.Errorf("failed to get setting value: %v", err)
+	}
+	if !settingValue {
+		return fmt.Errorf("expected setting value to be true, got false")
 	}
 
 	// queries can be cancelled using the context
