@@ -20,13 +20,14 @@ package std
 import (
 	"context"
 	"fmt"
-	clickhouse_tests "github.com/ClickHouse/clickhouse-go/v2/tests"
-	"github.com/stretchr/testify/require"
 	"math/rand"
 	"os"
 	"strconv"
 	"testing"
 	"time"
+
+	clickhouse_tests "github.com/ClickHouse/clickhouse-go/v2/tests"
+	"github.com/stretchr/testify/require"
 )
 
 func TestMain(m *testing.M) {
@@ -111,10 +112,6 @@ func TestStdMapInsertRead(t *testing.T) {
 	require.NoError(t, MapInsertRead())
 }
 
-func TestStdJSON(t *testing.T) {
-	require.NoError(t, JSONInsertRead())
-}
-
 func TestStdCompression(t *testing.T) {
 	require.NoError(t, CompressOpenDB())
 	require.NoError(t, CompressOpen())
@@ -130,10 +127,6 @@ func TestStdContext(t *testing.T) {
 
 func TestStdProgress(t *testing.T) {
 	require.NoError(t, ProgressProfileLogs())
-}
-
-func TestStdSession(t *testing.T) {
-	require.NoError(t, Sessions())
 }
 
 func TestStdDynamicScan(t *testing.T) {
