@@ -136,6 +136,14 @@ func (col *Array) Append(v any) (nulls []uint8, err error) {
 }
 
 func (col *Array) AppendRow(v any) error {
+	if col.depth == 1 {
+		// try to use reflection-free method.
+		return col.appendRowPlain(v)
+	}
+	return col.appendRowDefault(v)
+}
+
+func (col *Array) appendRowDefault(v any) error {
 	var elem reflect.Value
 	switch v := v.(type) {
 	case reflect.Value:
@@ -158,24 +166,73 @@ func (col *Array) AppendRow(v any) error {
 	return col.append(elem, 0)
 }
 
+func appendRowPlain[T any](col *Array, arr []T) error {
+	col.appendOffset(0, uint64(len(arr)))
+	for _, item := range arr {
+		if err := col.values.AppendRow(item); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func appendNullableRowPlain[T any](col *Array, arr []*T) error {
+	col.appendOffset(0, uint64(len(arr)))
+	for _, item := range arr {
+		var err error
+		if item == nil {
+			err = col.values.AppendRow(nil)
+		} else {
+			err = col.values.AppendRow(item)
+		}
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 func (col *Array) append(elem reflect.Value, level int) error {
 	if level < col.depth {
-		offset := uint64(elem.Len())
-		if ln := col.offsets[level].values.Rows(); ln != 0 {
-			offset += col.offsets[level].values.col.Row(ln - 1)
-		}
-		col.offsets[level].values.col.Append(offset)
-		for i := 0; i < elem.Len(); i++ {
-			if err := col.append(elem.Index(i), level+1); err != nil {
-				return err
+		switch elem.Kind() {
+		// reflect.Value.Len() & reflect.Value.Index() is called in `append` method which is only valid for
+		// Slice, Array and String that make sense here.
+		case reflect.Slice, reflect.Array, reflect.String:
+			col.appendOffset(level, uint64(elem.Len()))
+			for i := 0; i < elem.Len(); i++ {
+				el := elem.Index(i)
+
+				if el.Kind() == reflect.Interface && !el.IsNil() {
+					el = el.Elem()
+				}
+
+				if el.Kind() == reflect.Ptr && !el.IsNil() {
+					el = el.Elem()
+				}
+
+				if err := col.append(el, level+1); err != nil {
+					return err
+				}
 			}
+			return nil
 		}
-		return nil
+		return &ColumnConverterError{
+			Op:   "AppendRow",
+			To:   "Array",
+			From: fmt.Sprintf("%T", elem),
+		}
 	}
 	if elem.Kind() == reflect.Ptr && elem.IsNil() {
 		return col.values.AppendRow(nil)
 	}
 	return col.values.AppendRow(elem.Interface())
+}
+
+func (col *Array) appendOffset(level int, offset uint64) {
+	if ln := col.offsets[level].values.Rows(); ln != 0 {
+		offset += col.offsets[level].values.col.Row(ln - 1)
+	}
+	col.offsets[level].values.col.Append(offset)
 }
 
 func (col *Array) Decode(reader *proto.Reader, rows int) error {
