@@ -19,6 +19,7 @@ package column
 
 import (
 	"database/sql"
+	"database/sql/driver"
 	"fmt"
 	"github.com/ClickHouse/ch-go/proto"
 	"reflect"
@@ -64,7 +65,7 @@ func (col *Bool) ScanRow(dest any, row int) error {
 	case **bool:
 		*d = new(bool)
 		**d = col.row(row)
-	case *sql.NullBool:
+	case sql.Scanner:
 		return d.Scan(col.row(row))
 	default:
 		if scan, ok := dest.(sql.Scanner); ok {
@@ -113,6 +114,18 @@ func (col *Bool) Append(v any) (nulls []uint8, err error) {
 			col.Append(v[i])
 		}
 	default:
+		if valuer, ok := v.(driver.Valuer); ok {
+			val, err := valuer.Value()
+			if err != nil {
+				return nil, &ColumnConverterError{
+					Op:   "Append",
+					To:   "Bool",
+					From: fmt.Sprintf("%T", v),
+					Hint: "could not get driver.Valuer value",
+				}
+			}
+			return col.Append(val)
+		}
 		return nil, &ColumnConverterError{
 			Op:   "Append",
 			To:   "Bool",
@@ -143,6 +156,18 @@ func (col *Bool) AppendRow(v any) error {
 		}
 	case nil:
 	default:
+		if valuer, ok := v.(driver.Valuer); ok {
+			val, err := valuer.Value()
+			if err != nil {
+				return &ColumnConverterError{
+					Op:   "AppendRow",
+					To:   "Bool",
+					From: fmt.Sprintf("%T", v),
+					Hint: "could not get driver.Valuer value",
+				}
+			}
+			return col.AppendRow(val)
+		}
 		return &ColumnConverterError{
 			Op:   "AppendRow",
 			To:   "Bool",
