@@ -32,9 +32,12 @@ import (
 var (
 	//go:embed column.tpl
 	columnSrc string
+	//go:embed array.tpl
+	arraySrc string
 )
 var (
-	types []_type
+	types            []_type
+	supportedGoTypes []string
 )
 
 type _type struct {
@@ -65,6 +68,19 @@ func init() {
 	sort.Slice(types, func(i, j int) bool {
 		return sequenceKey(types[i].ChType) < sequenceKey(types[j].ChType)
 	})
+
+	for _, typ := range types {
+		supportedGoTypes = append(supportedGoTypes, typ.GoType)
+	}
+	supportedGoTypes = append(supportedGoTypes,
+		"string", "[]byte", "sql.NullString",
+		"int", "uint", "big.Int", "decimal.Decimal",
+		"bool", "sql.NullBool",
+		"time.Time", "sql.NullTime",
+		"uuid.UUID",
+		"netip.Addr", "net.IP", "proto.IPv6", "[16]byte",
+		"orb.MultiPolygon", "orb.Point", "orb.Polygon", "orb.Ring",
+	)
 }
 func write(name string, v any, t *template.Template) error {
 	out := new(bytes.Buffer)
@@ -87,10 +103,14 @@ func write(name string, v any, t *template.Template) error {
 }
 
 func main() {
-	for name, tpl := range map[string]*template.Template{
-		"column_gen": template.Must(template.New("column").Parse(columnSrc)),
+	for name, tpl := range map[string]struct {
+		template *template.Template
+		args     any
+	}{
+		"column_gen": {template.Must(template.New("column").Parse(columnSrc)), types},
+		"array_gen":  {template.Must(template.New("array").Parse(arraySrc)), supportedGoTypes},
 	} {
-		if err := write(name, types, tpl); err != nil {
+		if err := write(name, tpl.args, tpl.template); err != nil {
 			log.Fatal(err)
 		}
 	}
