@@ -19,15 +19,17 @@ package column
 
 import (
 	"database/sql"
+	"database/sql/driver"
 	"fmt"
-	"github.com/ClickHouse/ch-go/proto"
 	"reflect"
 	"time"
+
+	"github.com/ClickHouse/ch-go/proto"
 )
 
 var (
-	minDate32, _ = time.Parse("2006-01-02 15:04:05", "1925-01-01 00:00:00")
-	maxDate32, _ = time.Parse("2006-01-02 15:04:05", "2283-11-11 00:00:00")
+	minDate32, _ = time.Parse("2006-01-02 15:04:05", "1900-01-01 00:00:00")
+	maxDate32, _ = time.Parse("2006-01-02 15:04:05", "2299-12-31 00:00:00")
 )
 
 type Date32 struct {
@@ -90,9 +92,6 @@ func (col *Date32) Append(v any) (nulls []uint8, err error) {
 	switch v := v.(type) {
 	case []time.Time:
 		for _, t := range v {
-			if err := dateOverflow(minDate32, maxDate32, t, "2006-01-02"); err != nil {
-				return nil, err
-			}
 			col.col.Append(t)
 		}
 	case []*time.Time:
@@ -100,9 +99,6 @@ func (col *Date32) Append(v any) (nulls []uint8, err error) {
 		for i, v := range v {
 			switch {
 			case v != nil:
-				if err := dateOverflow(minDate32, maxDate32, *v, "2006-01-02"); err != nil {
-					return nil, err
-				}
 				col.col.Append(*v)
 			default:
 				nulls[i] = 1
@@ -146,6 +142,18 @@ func (col *Date32) Append(v any) (nulls []uint8, err error) {
 			}
 		}
 	default:
+		if valuer, ok := v.(driver.Valuer); ok {
+			val, err := valuer.Value()
+			if err != nil {
+				return nil, &ColumnConverterError{
+					Op:   "Append",
+					To:   "Date32",
+					From: fmt.Sprintf("%T", v),
+					Hint: "could not get driver.Valuer value",
+				}
+			}
+			return col.Append(val)
+		}
 		return nil, &ColumnConverterError{
 			Op:   "Append",
 			To:   "Date32",
@@ -158,16 +166,10 @@ func (col *Date32) Append(v any) (nulls []uint8, err error) {
 func (col *Date32) AppendRow(v any) error {
 	switch v := v.(type) {
 	case time.Time:
-		if err := dateOverflow(minDate32, maxDate32, v, "2006-01-02"); err != nil {
-			return err
-		}
 		col.col.Append(v)
 	case *time.Time:
 		switch {
 		case v != nil:
-			if err := dateOverflow(minDate32, maxDate32, *v, "2006-01-02"); err != nil {
-				return err
-			}
 			col.col.Append(*v)
 		default:
 			col.col.Append(time.Time{})
@@ -205,6 +207,18 @@ func (col *Date32) AppendRow(v any) error {
 			col.col.Append(value)
 		}
 	default:
+		if valuer, ok := v.(driver.Valuer); ok {
+			val, err := valuer.Value()
+			if err != nil {
+				return &ColumnConverterError{
+					Op:   "AppendRow",
+					To:   "Date32",
+					From: fmt.Sprintf("%T", v),
+					Hint: "could not get driver.Valuer value",
+				}
+			}
+			return col.AppendRow(val)
+		}
 		s, ok := v.(fmt.Stringer)
 		if ok {
 			return col.AppendRow(s.String())
@@ -233,10 +247,10 @@ func (col *Date32) Encode(buffer *proto.Buffer) {
 func (col *Date32) row(i int) time.Time {
 	t := col.col.Row(i)
 
-	if col.location != nil {
+	if col.location != nil && col.location != time.UTC {
 		// proto.Date is normalized as time.Time with UTC timezone.
 		// We make sure Date return from ClickHouse matches server timezone or user defined location.
-		t = time.Date(t.Year(), t.Month(), t.Day(), t.Hour(), t.Minute(), t.Second(), t.Nanosecond(), col.location)
+		t = getTimeWithDifferentLocation(t, col.location)
 	}
 	return t
 }
