@@ -22,8 +22,9 @@ import (
 	"database/sql/driver"
 	"encoding"
 	"fmt"
-	"github.com/ClickHouse/ch-go/proto"
 	"reflect"
+
+	"github.com/ClickHouse/ch-go/proto"
 
 	"github.com/ClickHouse/clickhouse-go/v2/lib/binary"
 )
@@ -77,7 +78,27 @@ func (col *FixedString) ScanRow(dest any, row int) error {
 		**d = col.row(row)
 	case encoding.BinaryUnmarshaler:
 		return d.UnmarshalBinary(col.rowBytes(row))
+	case *[]byte:
+		*d = col.rowBytes(row)
 	default:
+		// handle for *[n]byte
+		if t := reflect.TypeOf(dest); t.Kind() == reflect.Pointer &&
+			t.Elem().Kind() == reflect.Array &&
+			t.Elem().Elem() == reflect.TypeOf(byte(0)) {
+			size := t.Elem().Len()
+			if size != col.col.Size {
+				return &ColumnConverterError{
+					Op:   "ScanRow",
+					To:   fmt.Sprintf("%T", dest),
+					From: "FixedString",
+					Hint: fmt.Sprintf("invalid size %d, expect %d", size, col.col.Size),
+				}
+			}
+			rv := reflect.ValueOf(dest).Elem()
+			reflect.Copy(rv, reflect.ValueOf(col.row(row)))
+			return nil
+		}
+
 		if scan, ok := dest.(sql.Scanner); ok {
 			return scan.Scan(col.row(row))
 		}
@@ -125,7 +146,51 @@ func (col *FixedString) Append(v any) (nulls []uint8, err error) {
 		}
 		col.col.Append(data)
 		nulls = make([]uint8, len(data)/col.col.Size)
+	case [][]byte:
+		nulls = make([]uint8, len(v))
+		for i, v := range v {
+			if v == nil {
+				nulls[i] = 1
+			}
+			n := len(v)
+			if n == 0 {
+				col.col.Append(make([]byte, col.col.Size))
+			} else if n >= col.col.Size {
+				col.col.Append(v[0:col.col.Size])
+			} else {
+				data := make([]byte, col.col.Size)
+				copy(data, v)
+				col.col.Append(data)
+			}
+		}
 	default:
+		// handle for [][n]byte
+		if t := reflect.TypeOf(v); t.Kind() == reflect.Slice &&
+			t.Elem().Kind() == reflect.Array &&
+			t.Elem().Elem() == reflect.TypeOf(byte(0)) {
+			rv := reflect.ValueOf(v)
+			nulls = make([]uint8, rv.Len())
+			for i := 0; i < rv.Len(); i++ {
+				e := rv.Index(i)
+				data := make([]byte, e.Len())
+				reflect.Copy(reflect.ValueOf(data), e)
+				col.col.Append(data)
+			}
+			return
+		}
+
+		if s, ok := v.(driver.Valuer); ok {
+			val, err := s.Value()
+			if err != nil {
+				return nil, &ColumnConverterError{
+					Op:   "Append",
+					To:   "FixedString",
+					From: fmt.Sprintf("%T", s),
+					Hint: "could not get driver.Valuer value",
+				}
+			}
+			return col.Append(val)
+		}
 		return nil, &ColumnConverterError{
 			Op:   "Append",
 			To:   "FixedString",
@@ -138,6 +203,8 @@ func (col *FixedString) Append(v any) (nulls []uint8, err error) {
 func (col *FixedString) AppendRow(v any) (err error) {
 	data := make([]byte, col.col.Size)
 	switch v := v.(type) {
+	case []byte:
+		copy(data, v)
 	case string:
 		if v != "" {
 			data = binary.Str2Bytes(v, col.col.Size)
@@ -154,27 +221,31 @@ func (col *FixedString) AppendRow(v any) (err error) {
 			return err
 		}
 	default:
+		if t := reflect.TypeOf(v); t.Kind() == reflect.Array && t.Elem() == reflect.TypeOf(byte(0)) {
+			if t.Len() != col.col.Size {
+				return &ColumnConverterError{
+					Op:   "AppendRow",
+					To:   "FixedString",
+					From: fmt.Sprintf("%T", v),
+					Hint: fmt.Sprintf("invalid size %d, expect %d", t.Len(), col.col.Size),
+				}
+			}
+			reflect.Copy(reflect.ValueOf(data), reflect.ValueOf(v))
+			col.col.Append(data)
+			return nil
+		}
+
 		if s, ok := v.(driver.Valuer); ok {
 			val, err := s.Value()
 			if err != nil {
 				return &ColumnConverterError{
 					Op:   "AppendRow",
-					To:   "String",
+					To:   "FixedString",
 					From: fmt.Sprintf("%T", s),
 					Hint: "could not get driver.Valuer value",
 				}
 			}
-
-			if s, ok := val.(string); ok {
-				return col.AppendRow(s)
-			}
-
-			return &ColumnConverterError{
-				Op:   "AppendRow",
-				To:   "String",
-				From: fmt.Sprintf("%T", v),
-				Hint: "driver.Valuer value is not a string",
-			}
+			return col.AppendRow(val)
 		}
 
 		if s, ok := v.(fmt.Stringer); ok {
@@ -183,7 +254,7 @@ func (col *FixedString) AppendRow(v any) (err error) {
 
 		return &ColumnConverterError{
 			Op:   "AppendRow",
-			To:   "String",
+			To:   "FixedString",
 			From: fmt.Sprintf("%T", v),
 		}
 	}
