@@ -18,6 +18,7 @@
 package column
 
 import (
+	"database/sql/driver"
 	"fmt"
 	"github.com/ClickHouse/ch-go/proto"
 	"reflect"
@@ -88,13 +89,30 @@ func (col *Point) Append(v any) (nulls []uint8, err error) {
 		}
 	case []*orb.Point:
 		nulls = make([]uint8, len(v))
-		for _, v := range v {
-			col.col.Append(proto.Point{
-				X: v.Lon(),
-				Y: v.Lat(),
-			})
+		for i, v := range v {
+			if v == nil {
+				nulls[i] = 1
+				col.col.Append(proto.Point{})
+			} else {
+				col.col.Append(proto.Point{
+					X: v.Lon(),
+					Y: v.Lat(),
+				})
+			}
 		}
 	default:
+		if valuer, ok := v.(driver.Valuer); ok {
+			val, err := valuer.Value()
+			if err != nil {
+				return nil, &ColumnConverterError{
+					Op:   "Append",
+					To:   "Point",
+					From: fmt.Sprintf("%T", v),
+					Hint: fmt.Sprintf("could not get driver.Valuer value, try using %s", col.Type()),
+				}
+			}
+			return col.Append(val)
+		}
 		return nil, &ColumnConverterError{
 			Op:   "Append",
 			To:   "Point",
@@ -116,6 +134,18 @@ func (col *Point) AppendRow(v any) error {
 			Y: v.Lat(),
 		})
 	default:
+		if valuer, ok := v.(driver.Valuer); ok {
+			val, err := valuer.Value()
+			if err != nil {
+				return &ColumnConverterError{
+					Op:   "AppendRow",
+					To:   "Point",
+					From: fmt.Sprintf("%T", v),
+					Hint: fmt.Sprintf("could not get driver.Valuer value, try using %s", col.Type()),
+				}
+			}
+			return col.AppendRow(val)
+		}
 		return &ColumnConverterError{
 			Op:   "AppendRow",
 			To:   "Point",
