@@ -19,11 +19,21 @@ package column
 
 import (
 	"fmt"
-	"github.com/ClickHouse/ch-go/proto"
 	"reflect"
 	"strings"
 	"time"
+
+	"github.com/ClickHouse/ch-go/proto"
 )
+
+// This JSON type implementation was done for an experimental Object('JSON') type:
+// https://clickhouse.com/docs/en/sql-reference/data-types/object-data-type
+// It's already deprecated in ClickHouse and will be removed in the future.
+// Since ClickHouse 24.8, the Object('JSON') type is no longer alias for JSON type.
+// The new JSON type has been introduced: https://clickhouse.com/docs/en/sql-reference/data-types/newjson
+// However, the new JSON type is not supported by the driver yet.
+//
+// This implementation is kept for backward compatibility and will be removed in the future. TODO: remove this
 
 // inverse mapping - go types to clickhouse types
 var kindMappings = map[reflect.Kind]string{
@@ -40,7 +50,7 @@ var kindMappings = map[reflect.Kind]string{
 	reflect.Uint64:  "UInt64",
 	reflect.Float32: "Float32",
 	reflect.Float64: "Float64",
-	reflect.Bool:    "Boolean",
+	reflect.Bool:    "Bool",
 }
 
 // complex types for which a mapping exists - currently we map to String but could enhance in the future for other types
@@ -500,6 +510,13 @@ func appendStructOrMap(jCol *JSONObject, data any) error {
 				ColumnType: fmt.Sprint(reflect.TypeOf(data).Key().Kind()),
 				Err:        fmt.Errorf("map keys must be string for column %s", jCol.Name()),
 			}
+		}
+		if jCol.columns == nil && vData.Len() == 0 {
+			// if map is empty, we need to create an empty Tuple to make sure subcolumns protocol is happy
+			// _dummy is a ClickHouse internal name for empty Tuple subcolumn
+			// it has the same effect as `INSERT INTO single_json_type_table VALUES ('{}');`
+			jCol.upsertValue("_dummy", "Int8")
+			return jCol.insertEmptyColumn("_dummy")
 		}
 		return iterateMap(vData, jCol, 0)
 	}
