@@ -19,8 +19,11 @@ package tests
 
 import (
 	"context"
-	"github.com/stretchr/testify/require"
+	"database/sql/driver"
+	"fmt"
 	"testing"
+
+	"github.com/stretchr/testify/require"
 
 	"github.com/ClickHouse/clickhouse-go/v2"
 	"github.com/stretchr/testify/assert"
@@ -49,6 +52,7 @@ func TestSimpleEnum(t *testing.T) {
 	require.NoError(t, batch.Append(
 		col1Data,
 	))
+	require.Equal(t, 1, batch.Rows())
 	require.NoError(t, batch.Send())
 	var (
 		col1 string
@@ -86,6 +90,7 @@ func TestCustomEnum(t *testing.T) {
 		col2Data,
 		col3Data,
 	))
+	require.Equal(t, 1, batch.Rows())
 	require.NoError(t, batch.Send())
 	var (
 		col1 customStr
@@ -106,7 +111,7 @@ func TestEnum(t *testing.T) {
 	require.NoError(t, err)
 	const ddl = `
 			CREATE TABLE test_enum (
-				  Col1 Enum  ('hello'   = 1,  'world' = 2)
+				  Col1 Enum  ('hello,hello'   = 1,  'world' = 2)
 				, Col2 Enum8 ('click'   = 5,  'house' = 25)
 				, Col3 Enum16('house' = 10,   'value' = 50)
 				, Col4 Array(Enum8  ('click' = 1, 'house' = 2))
@@ -122,7 +127,7 @@ func TestEnum(t *testing.T) {
 	batch, err := conn.PrepareBatch(ctx, "INSERT INTO test_enum")
 	require.NoError(t, err)
 	var (
-		col1Data = "hello"
+		col1Data = "hello,hello"
 		col2Data = "click"
 		col3Data = "house"
 		col4Data = []string{"click", "house"}
@@ -139,6 +144,7 @@ func TestEnum(t *testing.T) {
 		col6Data,
 		col7Data,
 	))
+	require.Equal(t, 1, batch.Rows())
 	require.NoError(t, batch.Send())
 	var (
 		col1 string
@@ -182,6 +188,7 @@ func TestNullableEnum(t *testing.T) {
 	batch, err := conn.PrepareBatch(ctx, "INSERT INTO test_enum")
 	require.NoError(t, err)
 	require.NoError(t, batch.Append("hello", "click", "value"))
+	require.Equal(t, 1, batch.Rows())
 	require.NoError(t, batch.Send())
 	var (
 		col1 string
@@ -196,6 +203,7 @@ func TestNullableEnum(t *testing.T) {
 	batch, err = conn.PrepareBatch(ctx, "INSERT INTO test_enum")
 	require.NoError(t, err)
 	require.NoError(t, batch.Append("hello", nil, "value"))
+	require.Equal(t, 1, batch.Rows())
 	require.NoError(t, batch.Send())
 	{
 		var (
@@ -264,6 +272,7 @@ func TestColumnarEnum(t *testing.T) {
 	require.NoError(t, batch.Column(6).Append([][]*string{
 		col7Data, col7Data, col7Data,
 	}))
+	require.Equal(t, 3, batch.Rows())
 	require.NoError(t, batch.Send())
 	var (
 		col1 string
@@ -275,6 +284,87 @@ func TestColumnarEnum(t *testing.T) {
 		col7 []*string
 	)
 	require.NoError(t, conn.QueryRow(ctx, "SELECT * FROM test_enum LIMIT 1").Scan(
+		&col1, &col2, &col3, &col4,
+		&col5, &col6, &col7,
+	))
+	assert.Equal(t, col1Data, col1)
+	assert.Equal(t, col2Data, col2)
+	assert.Equal(t, col3Data, col3)
+	assert.Equal(t, col4Data, col4)
+	assert.Equal(t, col5Data, col5)
+	assert.Equal(t, col6Data, col6)
+	assert.Equal(t, col7Data, col7)
+}
+
+type testEnumSerializer struct {
+	val []string
+}
+
+func (c testEnumSerializer) Value() (driver.Value, error) {
+	return c.val, nil
+}
+
+func (c *testEnumSerializer) Scan(src any) error {
+	if t, ok := src.([]string); ok {
+		*c = testEnumSerializer{val: t}
+		return nil
+	}
+	return fmt.Errorf("cannot scan %T into testEnum8Serializer", src)
+}
+
+func TestEnumValuer(t *testing.T) {
+	conn, err := GetNativeConnection(nil, nil, &clickhouse.Compression{
+		Method: clickhouse.CompressionLZ4,
+	})
+	ctx := context.Background()
+	require.NoError(t, err)
+	const ddl = `
+			CREATE TABLE test_enum_valuer (
+				  Col1 Enum  ('hello'   = 1,  'world' = 2)
+				, Col2 Enum8 ('click'   = 5,  'house' = 25)
+				, Col3 Enum16('house' = 10,   'value' = 50)
+				, Col4 Array(Enum8  ('click' = 1, 'house' = 2))
+				, Col5 Array(Enum16 ('click' = 1, 'house' = 2))
+				, Col6 Array(Nullable(Enum8  ('click' = 1, 'house' = 2)))
+				, Col7 Array(Nullable(Enum16 ('click' = 1, 'house' = 2)))
+			) Engine MergeTree() ORDER BY tuple()
+		`
+	defer func() {
+		conn.Exec(ctx, "DROP TABLE IF EXISTS test_enum_valuer")
+	}()
+	require.NoError(t, conn.Exec(ctx, ddl))
+	batch, err := conn.PrepareBatch(ctx, "INSERT INTO test_enum_valuer")
+	require.NoError(t, err)
+	var (
+		col1Data = "hello"
+		col2Data = "click"
+		col3Data = "house"
+		col4Data = []string{"click", "house"}
+		col5Data = []string{"house", "click"}
+		col6Data = []*string{&col2Data, nil, &col3Data}
+		col7Data = []*string{&col3Data, nil, &col2Data}
+	)
+	require.NoError(t, batch.Append(
+		col1Data,
+		col2Data,
+		col3Data,
+		testEnumSerializer{val: col4Data},
+		testEnumSerializer{val: col5Data},
+		col6Data,
+		col7Data,
+	))
+	require.Equal(t, 1, batch.Rows())
+	require.NoError(t, batch.Send())
+	var (
+		col1 string
+		col2 string
+		col3 string
+		col4 []string
+		col5 []string
+		col6 []*string
+		col7 []*string
+	)
+	require.NoError(t, conn.QueryRow(ctx, "SELECT * FROM test_enum_valuer").Scan(
 		&col1, &col2, &col3, &col4,
 		&col5, &col6, &col7,
 	))
