@@ -19,20 +19,22 @@ package issues
 
 import (
 	"fmt"
+	"net/url"
+	"strconv"
+	"testing"
+
 	"github.com/ClickHouse/clickhouse-go/v2"
 	clickhouse_tests "github.com/ClickHouse/clickhouse-go/v2/tests"
 	"github.com/ClickHouse/clickhouse-go/v2/tests/std"
 	clickhouse_std_tests "github.com/ClickHouse/clickhouse-go/v2/tests/std"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"strconv"
-	"testing"
 )
 
 func TestIssue692(t *testing.T) {
 	useSSL, err := strconv.ParseBool(clickhouse_tests.GetEnv("CLICKHOUSE_USE_SSL", "false"))
 	require.NoError(t, err)
-	conn, err := clickhouse_std_tests.GetDSNConnection("issues", clickhouse.Native, useSSL, nil)
+	conn, err := clickhouse_std_tests.GetDSNConnection("issues", clickhouse.Native, useSSL, url.Values{})
 	require.NoError(t, err)
 	if !std.CheckMinServerVersion(conn, 21, 9, 0) {
 		t.Skip(fmt.Errorf("unsupported clickhouse version"))
@@ -44,7 +46,7 @@ func TestIssue692(t *testing.T) {
 			, Col2 Map(String, UInt64)
 			, Col3 Map(String, UInt64)
 			, Col4 Array(Map(String, String))
-			, Col5 Map(LowCardinality(String), LowCardinality(UInt64))
+			, Col5 Map(LowCardinality(String), LowCardinality(String))
 			, Col6 Map(String, Map(String, Int64))
 		) Engine MergeTree() ORDER BY tuple()
 		`
@@ -73,9 +75,9 @@ func TestIssue692(t *testing.T) {
 			{"A": "B"},
 			{"C": "D"},
 		}
-		col5Data = map[string]uint64{
-			"key_col_5_1": 100,
-			"key_col_5_2": 200,
+		col5Data = map[string]string{
+			"key_col_5_1": "100",
+			"key_col_5_2": "200",
 		}
 		col6Data = map[string]map[string]int64{}
 	)
@@ -87,7 +89,7 @@ func TestIssue692(t *testing.T) {
 		col2 map[string]uint64
 		col3 map[string]uint64
 		col4 []map[string]string
-		col5 map[string]uint64
+		col5 map[string]string
 		col6 map[string]map[string]int64
 	)
 	require.NoError(t, conn.QueryRow("SELECT Col1, Col2, Col3, Col4, Col5, Col6 FROM test_map").Scan(&col1, &col2, &col3, &col4, &col5, &col6))
